+++ database/seeders/BooksTableSeeder.php
@@ -2,7 +2,7 @@
 
 namespace Database\Seeders;
 
-use App\Book;
+use App\Models\Book;
 use Illuminate\Database\Seeder;
 
 class BooksTableSeeder extends Seeder
